+++ library_jpa/src/main/java/ru/otus/homework6/dao/StyleRepository.java
@@ -5,7 +5,7 @@ import ru.otus.homework6.domain.Style;
 import java.util.Optional;
 
 public interface StyleRepository {
-    Style getStyleById(long styleId);
+    Style getStyleById(int styleId);
     Optional<Style> getStyleByName(String styleName);
-    long saveNewStyle(Style style);
+    int saveNewStyle(Style style);
 }
